Rename file to match struct name
